Fire the BreakBlockEvent on the client and server
