+++ src/main/java/net/neoforged/neoforge/common/CommonHooks.java
@@ -211,6 +211,7 @@ import net.neoforged.neoforge.event.entity.player.SweepAttackEvent;
 import net.neoforged.neoforge.event.level.BlockDropsEvent;
 import net.neoforged.neoforge.event.level.BlockEvent;
 import net.neoforged.neoforge.event.level.NoteBlockEvent;
+import net.neoforged.neoforge.event.level.block.BreakBlockEvent;
 import net.neoforged.neoforge.event.level.block.CropGrowEvent;
 import net.neoforged.neoforge.fluids.FluidType;
 import net.neoforged.neoforge.internal.NeoForgeProxy;
@@ -572,7 +573,7 @@ public class CommonHooks {
      * @param state    The state of the block being broken
      * @return The event
      */
-    public static BlockEvent.BreakEvent fireBlockBreak(Level level, GameType gameType, ServerPlayer player, BlockPos pos, BlockState state) {
+    public static BreakBlockEvent fireBlockBreak(Level level, GameType gameType, Player player, BlockPos pos, BlockState state) {
         boolean preCancelEvent = false;
 
         ItemStack itemstack = player.getMainHandItem();
@@ -588,14 +589,13 @@ public class CommonHooks {
             preCancelEvent = true;
         }
 
-        // Post the block break event
-        BlockEvent.BreakEvent event = new BlockEvent.BreakEvent(level, pos, state, player);
+        var event = new BreakBlockEvent(level, pos, state, player);
         event.setCanceled(preCancelEvent);
         NeoForge.EVENT_BUS.post(event);
 
-        // If the event is canceled, let the client know the block still exists
-        if (event.isCanceled()) {
-            player.connection.send(new ClientboundBlockUpdatePacket(pos, state));
+        // If the event is canceled on the server, let the client know the block still exists
+        if (event.isCanceled() && event.shouldNotifyClient() && player instanceof ServerPlayer sp) {
+            sp.connection.send(new ClientboundBlockUpdatePacket(pos, state));
         }
 
         return event;
